pyproject.toml
[project]
name = "alecrim-audio-to-text"
version = "0.1.0"
description = "Biblioteca para usar áudio para texto usando o Cognitive Services da Azure."
readme = "README.md"
authors = [
    { name = "eng-alecrim", email = "eng.alecrim@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "azure-cognitiveservices-speech>=1.37.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
  "black[jupyter]>=24.10.0",
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
  "ipykernel>=6.29.5",
  "ipywidgets>=8.1.5",
  "taskipy>=1.14.1",
  "ruff>=0.9.2",
  "ignr>=2.2",
  "requests>=2.32.3",
]

[tool.ruff]
line-length = 88
extend-exclude = ["migrations", "notebooks", "*.ipynb", "lab"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]
ignore = ["PLR1711", "E731"]

[tool.ruff.format]
preview = true
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix && ruff format ."
pre_test = "task lint"
test = "pytest tests -s -x --cov=packages --cov-report=html --cov-config=.coveragerc -vv"
post_test = "coverage html"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
